flight, flightPrice, airport ADMIN section
flightPrice and airport currently does not work :(

+++ AirTicketBookingSystem/AirTicketBookingSystem-Api/src/main/java/cz/fi/muni/pa036/airticketbooking/api/dto/AirlineDto.java
@@ -12,7 +12,7 @@ import java.util.Set;
 public class AirlineDto  implements java.io.Serializable {
 
 
-     private int id;
+     private Long id;
      private String name;
      private String code;
      private Character mainAirline;
@@ -23,10 +23,10 @@ public class AirlineDto  implements java.io.Serializable {
     }
 
 	
-    public AirlineDto(int id) {
+    public AirlineDto(Long id) {
         this.id = id;
     }
-    public AirlineDto(int id, String name, String code, Character mainAirline, Set administrators, Set planes) {
+    public AirlineDto(Long id, String name, String code, Character mainAirline, Set administrators, Set planes) {
        this.id = id;
        this.name = name;
        this.code = code;
@@ -35,11 +35,11 @@ public class AirlineDto  implements java.io.Serializable {
        this.planes = planes;
     }
    
-    public int getId() {
+    public Long getId() {
         return this.id;
     }
     
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
     public String getName() {
